The system counter needs to be enabled ahead of time
It now reflects the fact I'm using one of 16-bit timers for system accurate timing (well as best as possible)

+++ cortex/IODevice.cc
@@ -113,7 +113,7 @@ namespace cortex
                 return 0;
             }
             void
-            setCompareValue(uint8_t value) noexcept {
+            setCompareValue(uint16_t value) noexcept {
                 Opcode code(0, Devices::Timer, Opcodes::CompareValue, 0);
                 code.write16(value);
             }
